refactor useFirebaseContext; add inter font;
