+++ src/components/pages/Home.tsx
@@ -1,7 +1,7 @@
 import { app } from "../../FirebaseConfig";
 import { User, signOut } from "firebase/auth";
 import React, { useEffect, useState } from "react";
-import { useFirebaseContext } from "../../context/FirebaseContext";
+import useFirebaseContext from "@/hooks/useFirebaseContext";
 import Icon from '@mdi/react'
 import { mdiLogout } from '@mdi/js'
